Attach colliders to their containing bodies only if there are such bodies
This commit replaces the old implementation of
`attach_bodies_and_colliders_system` that wrongly assumed that the
parent of a bodyless collider always has an attachable body and caused
runtime panics in `rapier3d::pipeline::handle_user_changes_to_
colliders`, where it attempted to look up the non-existent bodies
attached to given colliders.

As with the old implementation, bodies can include colliders defined in
their children. In the new implementation, all non-direct descendants
are also eligible for inclusion. I think this behavior is more in
accordance with the users' expectation.

Fixes #76.