Add a pointer from `docs.rs` docs to rapier docs
Added a pointer from the module-level cargo docs to the actual docs on `rapier.rs` so other people won't be lost like I was.

(Searches for bevy-rapier3d documentation often end up on `docs.rs`, where the module appears to be almost completely undocumented. The Bevy docs also make little or no mention of it. The README has a pointer to the docs, but it never even occurred to me that that might go anywhere other than docs.rs.)