Add debug render mesh for capsule collider shape.
Without changing the way that the debug mesh's transform is calculated (just from scale), it's not
possible to accurately show a capsule's orientation in the debug view. So, just use a cuboid and
scale it along the capsule's AABB to get a reasonably accurate approximation.

I also did a minor tidy of the transform sync code.